Handled exceptions in readCharacteristics. Modified some compose functions in Views. Moved some files to new packages. Created packages for common composable functions
