--- app/src/main/java/com/untitledkingdom/ueberapp/ui/Colors.kt
@@ -1,62 +1,0 @@
-package com.untitledkingdom.ueberapp.ui
-
-import androidx.compose.ui.graphics.Color
-
-object Colors {
-    val Purple = Color(0xFF9E87F5)
-    val RoomNameGray = Color(0xFF5B5B5B)
-    val Gray10 = Color(0x1AA6A5A5)
-    val LightPurple = Color(0xFFD1D6FE)
-    val LightGray = Color(0xFFC6C6C8)
-    val EventOnline = Color(0xFFC6C6C8)
-    val MediumGray = Color(0xFFA6A6A5)
-    val SplashPurple = Color(0xFF9E87F5)
-    val UnhighlightedDeviceNameGray = Color(0xFF5B5B5B)
-    val UnHighlightedDeviceInfoGray = Color(0xFFA6A5A5)
-    val White10 = Color(0x1AFFFFFF)
-    val Blue = Color(0xFF1782FA)
-    val AppBackground = Color(0xFFF9F9FF)
-    val AppBackgroundBlack = Color(0xFF151A23)
-    val FilterBlue = Color(0xFF1782FA)
-    val GrayCardNotAvailable = Color(0xFFF9F9F9)
-    val GrayCardNotAvailableDark = Color(0x1AA6A5A5)
-    val GrayOsVersion = Color(0xFFA7A7A7)
-    val MyEventsCalendarDayBlack = Color(0xFF151A23)
-    val SectionDividerLight = Color(0xFFC6C6C8)
-    val SectionDividerDark = Color(0xFF5B5B5B)
-    val FilterRed = Color(0xFFFB1515)
-    val FilterGray = Color(0xFFA6A5A5)
-    val Gray = Color(0xFF909090)
-    val GrayCalendarItemText = Color(0xFF5B5B5B)
-    val Black = Color(0xFF000000)
-    val BlackSelectedDay = Color(0xFF151A23)
-    val Purple200 = Color(0xFFBB86FC)
-    val Purple500 = Color(0xFF6200EE)
-    val Purple700 = Color(0xFF3700B3)
-    val Teal200 = Color(0xFF03DAC5)
-    val ColorFilterBackgroundDark = Color(0x33D1D6FE)
-    val Toolbar = Color(0xFFFFFFFF)
-    val White = Color(0xFFFFFFFF)
-    val Divider = Color(0x14212121)
-    val DeviceNameDeviceDetailsColor = Color(0xFF151A23)
-    val DevicesTabsColorBlack = Color(0xFF282D39)
-    val DevicesTabsColorDark = Color(0x24D1D6FE)
-    val ButtonDialogColor = Color(0xFF1782FA)
-    val RoomName1Color = Color(0xFFC06591)
-    val RoomName2Color = Color(0xFF9D65C0)
-    val RoomName3Color = Color(0xFF656EC0)
-    val RoomName4Color = Color(0xFF65B0C0)
-    val RoomName5Color = Color(0xFF84CD94)
-    val RoomName6Color = Color(0xFFCCCD84)
-    val RoomName7Color = Color(0xFFCD9684)
-    val OnlineEventColor = Color(0xFF858585)
-    val Purple400 = Color(0xFF7B1FFF)
-    val TextField = Color.Transparent
-    val ButtonSelected = Purple500
-    val ButtonGrayNotSelected = Color(0xFFf1f3f4)
-    val SnackbarSuccess: Color = Color(0xFF4CAF50)
-    val SnackbarError: Color = Color(0xFFF44336)
-    val secondaryVariant: Color = Color(0xff9E87F5)
-    val TimePickerBackground: Color = Color(0xFFFAFAFA)
-    val IconClearNextField: Color = Color(0xFF9E87F5)
-}
