--- app/src/main/java/com/untitledkingdom/ueberapp/ui/Paddings.kt
@@ -1,12 +1,0 @@
-package com.untitledkingdom.ueberapp.ui
-
-import androidx.compose.ui.unit.dp
-
-object Paddings {
-    val padding2 = 2.dp
-    val padding8 = 8.dp
-    val padding12 = 12.dp
-    val padding16 = 16.dp
-    val padding24 = 24.dp
-    val padding72 = 72.dp
-}
