--- app/src/main/java/com/untitledkingdom/ueberapp/ui/Shapes.kt
@@ -1,8 +1,0 @@
-package com.untitledkingdom.ueberapp.ui
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.ui.unit.dp
-
-object Shapes {
-    val shape8 = RoundedCornerShape(8.dp)
-}
