+++ app/src/main/java/com/untitledkingdom/ueberapp/ui/values/Paddings.kt
@@ -1,0 +1,10 @@
+package com.untitledkingdom.ueberapp.ui.values
+
+import androidx.compose.ui.unit.dp
+
+val padding2 = 2.dp
+val padding8 = 8.dp
+val padding12 = 12.dp
+val padding16 = 16.dp
+val padding24 = 24.dp
+val padding72 = 72.dp
