+++ app/src/main/java/com/untitledkingdom/ueberapp/ui/values/Shapes.kt
@@ -1,0 +1,6 @@
+package com.untitledkingdom.ueberapp.ui.values
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.ui.unit.dp
+
+val shape8 = RoundedCornerShape(8.dp)
