+++ app/src/main/res/drawable/ic_baseline_signal_3.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24"
+    android:tint="?attr/colorControlNormal">
+  <path
+      android:fillColor="@android:color/white"
+      android:pathData="M17,4h3v16h-3zM5,14h3v6L5,20zM11,9h3v11h-3z"/>
+</vector>
